[Rename] 파일명에 paging 제거 및 사용하지 않는 클래스 제거
