--- app/src/main/java/com/oyj/kakaobook/data/SearchUiState.kt
@@ -1,10 +1,0 @@
-package com.oyj.kakaobook.data
-
-sealed class SearchUiState
-
-object Init : SearchUiState()
-object Loading : SearchUiState()
-object Empty : SearchUiState()
-data class Success(val bookList: List<BookItem>) : SearchUiState()
-data class Error(val message: String) : SearchUiState()
-
