--- app/src/main/java/com/oyj/kakaobook/model/BookModel.kt
@@ -1,9 +1,0 @@
-package com.oyj.kakaobook.model
-
-import com.oyj.domain.entity.Book
-import com.oyj.kakaobook.data.BookItem
-
-data class BookModel(
-    val bookItem: BookItem,
-    val book: Book
-)
