+++ app/src/main/java/com/oyj/kakaobook/ui/component/SearchResultView.kt
@@ -3,28 +3,48 @@ package com.oyj.kakaobook.ui.component
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.oyj.kakaobook.data.BookItem
-import com.oyj.kakaobook.ui.component.ComponentConstants.Padding
+import androidx.paging.compose.LazyPagingItems
+import com.oyj.domain.entity.Book
+import com.oyj.kakaobook.mapper.PresenterMapper.toBookItem
+import com.oyj.kakaobook.ui.component.ComponentConstants.LazyColumn.BOTTOM_PADDING
+import com.oyj.kakaobook.ui.component.ComponentConstants.LazyColumn.VERTICAL_SPACING
 
 @Composable
 fun SearchResultView(
     modifier: Modifier = Modifier,
-    bookList: List<BookItem>,
-    onClickBookmark: (String) -> Unit = {}
+    bookList: LazyPagingItems<Book>,
+    bookmarkedIsbnSet: Set<String>,
+    onClickBookmark: (Book) -> Unit = {},
+    onClickCard: (Book) -> Unit = {}
 ) {
     LazyColumn(
         modifier = modifier,
-        contentPadding = PaddingValues(Padding.LARGE),
-        verticalArrangement = Arrangement.spacedBy(Padding.MEDIUM)
+        contentPadding = PaddingValues(bottom = BOTTOM_PADDING),
+        verticalArrangement = Arrangement.spacedBy(VERTICAL_SPACING)
     ) {
-        items(bookList) { book ->
+        items(
+            count = bookList.itemCount,
+            key = { index -> bookList[index]?.isbn ?: index }
+        ) { index ->
+            val isBookmark = bookmarkedIsbnSet.contains(bookList[index]?.isbn)
+            val bookItem = bookList[index]?.toBookItem(isBookmark) ?: return@items
             BookItemCard(
-                book = book,
-                onClickBookmark = onClickBookmark
+                book = bookItem,
+                onClickBookmark = {
+                    val book = findBookByIsbn(bookList, bookItem.isbn) ?: return@BookItemCard
+                    onClickBookmark(book)
+                },
+                onClickCard = {
+                    val book = findBookByIsbn(bookList, bookItem.isbn) ?: return@BookItemCard
+                    onClickCard(book)
+                }
             )
         }
     }
+}
+
+fun findBookByIsbn(bookList: LazyPagingItems<Book>, isbn: String): Book? {
+    return bookList.itemSnapshotList.items.firstOrNull { it.isbn == isbn }
 }
