+++ domain/src/main/java/com/oyj/domain/usecase/GetBookListUseCase.kt
@@ -1,7 +1,7 @@
 package com.oyj.domain.usecase
 
+import androidx.paging.PagingData
 import com.oyj.domain.entity.Book
-import com.oyj.domain.entity.Result
 import com.oyj.domain.repository.BookRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
@@ -9,9 +9,10 @@ import javax.inject.Inject
 class GetBookListUseCase @Inject constructor(
     private val repository: BookRepository
 ) {
-    suspend operator fun invoke(query: String): Flow<Result<List<Book>>> {
+    suspend operator fun invoke(query: String, sortCriteria: String): Flow<PagingData<Book>> {
         return repository.getBookList(
-            query = query
+            query = query,
+            sortCriteria = sortCriteria
         )
     }
 }
